requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.0
tqdm==4.65.0
psutil==5.9.5
segment-anything==1.0
matplotlib==3.7.2
scikit-learn==1.3.0
scikit-image==0.21.0
requests==2.31.0
geojson==3.0.1
streamlit==1.28.0
ultralytics==8.0.196
folium==0.14.0

